bug fix: 'toDecimal' had not been supported 64bit

+++ src/toDecimal.cpp
@@ -4,7 +4,7 @@ ll toDecimal(const std::string& binary) {
 	size_t n = binary.size();
 	ll res = 0;
 	for (size_t i = 0; i < binary.size(); ++i) {
-		res += (binary[n - i - 1] - '0') << i;
+		res += ll(binary[n - i - 1] - '0') << i;
 	}
 	return res;
 }
